requirements.txt
openai~=0.27.8
python-dotenv~=1.0.0
Jinja2~=3.1.2
colorama~=0.4.6
PyYAML~=6.0
chardet